# docs/openapi/device-api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movie Inbox Device API",
    "version": "1.0.0",
    "description": "Contract for native devices. This is not the historical browser API under /api/."
  },
  "servers": [{"url": "https://{baseUrl}", "variables": {"baseUrl": {"default": "movie-inbox.example"}}}],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "operationId": "createDeviceSession",
        "security": [],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}},
        "responses": {"201": {"description": "Device session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DeviceSession"}}}}, "401": {"$ref": "#/components/responses/Unauthorized"}, "429": {"$ref": "#/components/responses/RateLimited"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {"deviceBearer": {"type": "http", "scheme": "bearer", "bearerFormat": "opaque"}},
...
    "responses": {
      "Unauthorized": {"description": "Missing, invalid, expired or revoked device session", "headers": {"WWW-Authenticate": {"schema": {"type": "string"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
      "RateLimited": {"description": "Login rate limit", "headers": {"Retry-After": {"schema": {"type": "integer", "minimum": 1}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    },
    "schemas": {
      "LoginRequest": {"type": "object", "additionalProperties": false, "required": ["username", "password", "device_name"], "properties": {"username": {"type": "string", "minLength": 3, "maxLength": 64}, "password": {"type": "string", "minLength": 1, "maxLength": 1024, "writeOnly": true}, "device_name": {"type": "string", "minLength": 1, "maxLength": 80}}},
...
      "DeviceSession": {"type": "object", "additionalProperties": false, "required": ["access_token", "refresh_token", "token_type", "access_expires_at", "refresh_expires_at", "identity"], "properties": {"access_token": {"type": "string", "minLength": 32, "writeOnly": true}, "refresh_token": {"type": "string", "minLength": 32, "writeOnly": true}, "token_type": {"const": "Bearer"}, "access_expires_at": {"type": "string", "format": "date-time"}, "refresh_expires_at": {"type": "string", "format": "date-time"}, "identity": {"$ref": "#/components/schemas/Identity"}}},
      "Identity": {"type": "object", "additionalProperties": false, "required": ["user", "catalog"], "properties": {"user": {"type": "object", "additionalProperties": false, "required": ["id", "username", "role"], "properties": {"id": {"type": "string"}, "username": {"type": "string"}, "role": {"enum": ["owner", "member"]}}}, "catalog": {"type": "object", "additionalProperties": false, "required": ["id", "name"], "properties": {"id": {"type": "string"}, "name": {"type": "string"}}}}},
...
      "Error": {"type": "object", "additionalProperties": false, "required": ["error"], "properties": {"error": {"type": "object", "additionalProperties": false, "required": ["code"], "properties": {"code": {"type": "string"}}}}}
    }
  },
  "security": [{"deviceBearer": []}]
}